requirements.txt
boto3
awscli
pillow
gradio